requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.1.0
transformers>=4.20.0
tokenizers>=0.12.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0
Pillow>=9.0.0
evm-cfg-builder>=0.1.0
nltk>=3.7
gensim>=4.2.0
py-solc-x>=1.12.0
pyevmasm>=0.2.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

